settings: move menu to Services
